django_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3
django-filter>=23.2
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0